Move rescoring models from cc-utils
